+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Coerce textual lat/long to numeric in latlon-spherical
+- Return crossTrackDistance / alongTrackDistance of 0 when 'this' point equals start point
 - Round UTM to nm rather than (erroneous) μm
 - Fix (rare) rounding error issue in intersection() (#71]